Extract reason tests into dedicated modules
